HCF-1122 Use BASH_SOURCE instead of GIT_ROOT to locate script dirs
GIT_ROOT doesn't work when the repo is a submodule (i.e. being built
as part of fissile).

GIT_ROOT is still being used for versioning. So when configgin is
being built as part of fissile, then the dist results will embed
the fissile version. This shouldn't really matter, as the version
isn't embedded inside the configgin software itself.