kube: determine bootstrapness via image version
Instead of configuring all nodes of index 0 as bootstrap node, instead
set a pod as bootstrap if there are no other pods with the same version.
This makes it easier to deal with upgrades in a kube environment (where
the last pod - not necessarily index 0 - will be upgraded first).

This also means that we can configure some of the roles to no longer be
indexed (because they only needed it for bootstrapping).