KubeLinkSpecs: look for role pods without job name in properties
In #74, we changed configgin's exported properties to have a separate
set per job; this was a breaking change that made upgrading a cluster
problematic as mysql-1 would be unable to find a ready mysql-0, and
ended up being stuck.  This make it accept pods that either have the
job-specific properties, or the generic properties.