[WIP] Restart pods on imported property change
Currently just have code to mark property hashes on export and copy the hash to
the pods on import.  The hashing code is broken (it always hashes "null" instead
of the correct value).

Once this is done, we need to (from the pod that does the exporting) set the
same hashes on statefulsets which import those properties, under the
.spec.metadata.annotations key, such that kube will restart pods as appropriate.