CAP-761 Store exported properties in secrets instead of annotations
The exported secrets may contain credentials, so should not be
available to everyone able to get the pod details.

This change creates a separate secret for each pod, named using
both the pod name and uid (the name is purely for human convenience
and debugging). The secret uncludes an owner reference bacl to
the pod, so it will get removed whenever the pod is deleted.

The secret is only written for the main container of an instance
group, but not for the loggregator-agent sidecar. Therefore this
commit depends on an updated fissile release that sets the
KUBERNETES_CONTAINER_NAME environment variable correspondingly.

[jsc#CAP-761]