Avoid unnecessary pod restarts
Exported properties are no longer stored in secrets per pod, but in
a secret per stateful set. It records the exported values. It also
stores the CONFIGGIN_VERSION_TAG, which includes the chart version
and the secrets generation, so it will change whenever there is a
full restart of all pods.

Fissile will generate CONFIGGIN_IMPORT_$ROLE environment variables
that reference the CONFIGGIN_VERSION_TAG in the secret of the exporting
role. This prevents the importing pod from starting until the secret
with the exported properties has been written.

This obviously requires that any mutual or cyclical imports between
roles are resolved in the role manifest, either by using properties,
or by patching some scripts to remove the bosh link and break the
cycle. Otherwise there would be a deadlock as all roles in the cycle
would wait forever for their imports to become ready.

There is no problem with cycles in imports from different jobs inside
the same pod, as they are resolved simultaneously.

When properties are changed without a chart upgrade or secrets rotation,
the pod templates of all importing stateful sets are annotated with
the digest values of the exported values. The stateful set will
automatically restart pods when the checksum of the pod spec has changed.

To avoid redundant restarts when the properties match the original deployment
values, the annotations are not set (or removed if they already exist)
when the digest matches the initial digest value.

[jsc#CAP-150]