Allow service account override for configgin
If CONFIGGIN_SA_TOKEN is defined, use that value instead of the
pod service account for configgin. This is preferable so that the
rest of the pod does not have access to all the secrets in the
namespace.

The token may expire, but configgin only runs briefly when the pod
is created, so it should not be a problem that the token cannot be
refreshed.

[jsc#CAP-828]